Improves logging integration and stability
Adds metadata to track the source of QSOs, preventing feedback loops with logging software.

Removes the deprecated bridge functionality and introduces a debug endpoint to facilitate testing and troubleshooting of logging software integrations.

Refactors the logger integration to be more reliable and provide more detailed debugging information. Updates copy to be more accurate in admin panel.